Added fragments result activity
Both result list and result statistics are now seen in this activity. The statistics are displayed in a scrollview so that it is scalable for more metrics later on
