+++ app/src/main/res/layout/montly_calendar_cell.xml
@@ -28,7 +28,7 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintVertical_bias="0.6"
+        app:layout_constraintVertical_bias="0.75"
         android:visibility="invisible"
         tools:ignore="ContentDescription" />
